requirements.txt
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc